+++ src/DonorProfiles/resources/js/app/components/heading/index.js
@@ -1,0 +1,13 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+import './style.scss';
+
+const Heading = ( { icon, children } ) => {
+	return (
+		<div className="give-donor-profile-heading">
+			{ icon && ( <FontAwesomeIcon icon={ icon } /> ) }
+			{ children }
+		</div>
+	);
+};
+export default Heading;