+++ src/DonorProfiles/resources/js/app/index.js
@@ -3,13 +3,39 @@
 // Vendor dependencies
 import { HashRouter as Router } from 'react-router-dom';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
 
-// Reports app
-import App from '../components/app';
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+
+library.add( fas );
+
+// Store dependencies
+import { store } from './store';
+
+// Internal dependencies
+import { registerDefaultTabs } from './tabs';
+import { registerTab } from './utils';
+
+// DonorProfiles app
+import App from './components/app';
+
+import './style.scss';
+
+window.giveDonorProfile = {
+	store,
+	utils: {
+		registerTab,
+	},
+};
 
 ReactDOM.render(
-	<Router>
-		<App />
-	</Router>,
+	<Provider store={ store }>
+		<Router>
+			<App />
+		</Router>
+	</Provider>,
 	document.getElementById( 'give-donor-profile' )
 );
+
+registerDefaultTabs();