+++ src/DonorProfiles/resources/js/app/store/actions.js
@@ -1,0 +1,8 @@
+export const addTab = ( tab ) => {
+	return {
+		type: 'ADD_TAB',
+		payload: {
+			tab,
+		},
+	};
+};