+++ src/DonorProfiles/resources/js/app/tabs/annual-receipts/content.js
@@ -1,0 +1,13 @@
+import Heading from '../../components/heading';
+import { Fragment } from 'react';
+
+const Content = () => {
+	return (
+		<Fragment>
+			<Heading>
+				Annual Receipt
+			</Heading>
+		</Fragment>
+	);
+};
+export default Content;