+++ src/DonorProfiles/resources/js/app/tabs/dashboard/content.js
@@ -1,0 +1,20 @@
+import Heading from '../../components/heading';
+import RESTDonationTable from '../../components/rest-donation-table';
+import Stats from './stats';
+import { Fragment } from 'react';
+
+const Content = () => {
+	return (
+		<Fragment>
+			<Heading icon="chart-line">
+				Your Giving Stats
+			</Heading>
+			<Stats />
+			<Heading icon="calendar-alt">
+				Recent Donations
+			</Heading>
+			<RESTDonationTable />
+		</Fragment>
+	);
+};
+export default Content;