+++ src/DonorProfiles/resources/js/app/tabs/edit-profile/index.js
@@ -1,0 +1,9 @@
+// Internal dependencies
+import Content from './content';
+
+export const registerEditProfileTab = () => window.giveDonorProfile.utils.registerTab( {
+	label: 'Edit Profile',
+	icon: 'cog',
+	slug: 'edit-profile',
+	content: Content,
+} );