+++ src/DonorProfiles/resources/js/app/utils/index.js
@@ -1,0 +1,6 @@
+import { addTab } from '../store/actions';
+
+export const registerTab = ( tab ) => {
+	const { dispatch } = window.giveDonorProfile.store;
+	dispatch( addTab( tab ) );
+};