+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Front-end donor profiles are scaffolded (#5441)
+-   Donor Profiles UI is implemented (#5444)
 
 ### Fixed